requirements/base.txt
aws-cdk.core==1.52.0
aws-cdk.aws-lambda==1.52.0
-r requirements-deps.txt